# Optimize tsm1.FloatDecoder