## Overview

This commit changes the `FloatDecoder.val` from a `float64` type to a `uint64` to avoid an additional type conversion during read. Now the type gets converted to a `float64` only on call to `Values()`.

/cc @jwilder 

### Benchmark

```
benchmark                                   old ns/op     new ns/op     delta
BenchmarkEngine_CreateIterator_Count_1M     12145578      11839383      -2.52%

benchmark                                   old allocs     new allocs     delta
BenchmarkEngine_CreateIterator_Count_1M     74             73             -1.35%
```

###### Required for all non-trivial PRs
- [x] Rebased/mergable
- [x] Tests pass
- [x] Sign [CLA](https://influxdata.com/community/cla/) (if not already signed)