@jwilder 2016-04-20
```
[ERROR] run: Command 'go test -v -parallel 1 -timeout 480s ./...' failed with error: # github.com/influxdata/influxdb/tsdb/engine/tsm1
tsdb/engine/tsm1/engine_test.go:22:2: cannot find package "github.com/pkg/profile" in any of:
	/usr/local/go/src/github.com/pkg/profile (from $GOROOT)
	/root/go/src/github.com/pkg/profile (from $GOPATH)
```